[add, refactor, test] Testing typechecking for variables
Common lisp is really fun, planning to add this to the functions too!

Also I tried to setup asdf to make compilation, packaging and stuff
easier, like keeping types in a `type.lip` file, functions for getting
input in another, etc. But still no luck, have to research it a bit
more.